1.21.4 Working version
This is without a doubt NOT a foolproof update. However, it is a 'working' version. I am able to open links (tested with Hey Wiki). Able to reload, make new tabs, close tabs, go back a page.

Also the fps limiter is not working since i was not able to figure out how that worked just yet.